refactor: small changes, mainly with importing files (CAS-78)
