+++ app/models.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column, Integer, String, Boolean
-from .database import Base
+from database import Base
 
 
 class User(Base):
@@ -8,5 +8,5 @@ class User(Base):
     id = Column(Integer, primary_key=True, index=True)
     password = Column(String)
     email = Column(String, unique=True, index=True)
-    is_activee = Column(Boolean, default=False)
+    is_active = Column(Boolean, default=False)
     is_superuser = Column(Boolean, default=False)
