+++ requirements.txt
@@ -1,46 +1,60 @@
 alembic==1.13.1
-annotated-types==0.6.0
-anyio==4.2.0
-bcrypt==4.1.2
-certifi==2023.11.17
-cffi==1.16.0
-click==8.1.7
-cryptography==41.0.7
-dnspython==2.4.2
-ecdsa==0.18.0
-email-validator==2.1.0.post1
-exceptiongroup==1.2.0
+databases==0.8.0
 fastapi==0.108.0
-h11==0.14.0
-httpcore==1.0.2
-httptools==0.6.1
-httpx==0.26.0
-idna==3.6
-itsdangerous==2.1.2
-Jinja2==3.1.2
-Mako==1.3.0
-MarkupSafe==2.1.3
-orjson==3.9.10
+httpx==0.26.0 # testy, chyba niepotrzebne
 passlib==1.7.4
 psycopg2-binary==2.9.9
-pyasn1==0.5.1
-pycparser==2.21
-pydantic==2.5.3
-pydantic-extra-types==2.3.0
-pydantic-settings==2.1.0
-pydantic_core==2.14.6
-python-dotenv==1.0.0
-python-jose==3.3.0
+pytest==7.4.4 # testy
+# python-dotenv==1.0.0
+python-jose==3.3.0 # zastąpić za pomocą pyjwt
 python-multipart==0.0.6
-PyYAML==6.0.1
-rsa==4.9
-six==1.16.0
-sniffio==1.3.0
-SQLAlchemy==2.0.24
-starlette==0.32.0.post1
-typing_extensions==4.9.0
-ujson==5.9.0
+sqlalchemy==1.4.50 #2.0.24
 uvicorn==0.25.0
-uvloop==0.19.0
-watchfiles==0.21.0
-websockets==12.0
+
+
+# alembic==1.13.1
+# annotated-types==0.6.0
+# anyio==4.2.0
+# bcrypt==4.1.2
+# certifi==2023.11.17
+# cffi==1.16.0
+# click==8.1.7
+# cryptography==41.0.7
+# dnspython==2.4.2
+# ecdsa==0.18.0
+# email-validator==2.1.0.post1
+# exceptiongroup==1.2.0
+# fastapi==0.108.0
+# h11==0.14.0
+# httpcore==1.0.2
+# httptools==0.6.1
+# httpx==0.26.0
+# idna==3.6
+# itsdangerous==2.1.2
+# Jinja2==3.1.2
+# Mako==1.3.0
+# MarkupSafe==2.1.3
+# orjson==3.9.10
+# passlib==1.7.4
+# psycopg2-binary==2.9.9
+# pyasn1==0.5.1
+# pycparser==2.21
+# pydantic==2.5.3
+# pydantic-extra-types==2.3.0
+# pydantic-settings==2.1.0
+# pydantic_core==2.14.6
+# python-dotenv==1.0.0
+# python-jose==3.3.0
+# python-multipart==0.0.6
+# PyYAML==6.0.1
+# rsa==4.9
+# six==1.16.0
+# sniffio==1.3.0
+# SQLAlchemy==2.0.24
+# starlette==0.32.0.post1
+# typing_extensions==4.9.0
+# ujson==5.9.0
+# uvicorn==0.25.0
+# uvloop==0.19.0
+# watchfiles==0.21.0
+# websockets==12.0
